refactor: simplify implementation AYC-000
