+++ ayunis-core-frontend/src/pages/chat/ui/ChatPage.tsx
@@ -1,6 +1,7 @@
 import { useState, useRef, useEffect, useCallback, useMemo } from 'react';
 import ChatInterfaceLayout from '@/layouts/chat-interface-layout/ui/ChatInterfaceLayout';
 import ChatMessage from '@/pages/chat/ui/ChatMessage';
+import StreamingLoadingIndicator from '@/pages/chat/ui/StreamingLoadingIndicator';
 import ChatInput from '@/widgets/chat-input';
 import { useChatContext } from '@/shared/contexts/chat/useChatContext';
 import { useMessageSend } from '../api/useMessageSend';
@@ -500,6 +501,11 @@ export default function ChatPage({
   );
 
   // Chat Content (Messages only)
+  // Show loading indicator when streaming but assistant message hasn't arrived yet
+  const lastMessage = sortedMessages[sortedMessages.length - 1];
+  const showLoadingMessage =
+    isStreaming && (!lastMessage || lastMessage.role !== 'assistant');
+
   const chatContent = (
     <div className="p-4 pb-8">
       {sortedMessages.map((message, i) => (
@@ -514,6 +520,7 @@ export default function ChatPage({
           }
         />
       ))}
+      {showLoadingMessage && <StreamingLoadingIndicator />}
     </div>
   );
